# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: PromoGo API
  version: "1.0"
  description: >
    PromoGo's canonical HTTP API contract: health/readiness, the store-scoped
    1C/POS webhook API (accrual, redemption, client lookup, balance), the
    mobile-customer OTP/session API, and the retailer-staff/platform-admin
    OIDC + RBAC API. Every operation here corresponds 1:1 to a registered
    route in internal/httpserver's route table (see
    internal/httpserver/routes.go) — this parity is enforced by
    TestRouteOpenAPIParity.
  license:
    name: Proprietary
    url: https://github.com/MirzaDgtu/PromoGo
servers:
  - url: /
    description: Relative to wherever the API is deployed; there is no fixed public host.
security:
  - {}
...
paths:
...
  /api/v1/clients/resolve-qr:
    post:
      operationId: resolveQR
      tags: [Store]
      summary: Exchange a customer-issued QR payload for a store-scoped client (POS)
      description: >
        Atomically consumes a one-time QR payload minted by POST /me/qr,
        resolving (and creating/linking if this is the client's first visit
        to this store) a store-scoped client. Requires the `clients.lookup`
        scope. See DEC-011.
      security:
        - storeApiKey: [clients.lookup]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResolveQRRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ResolveQRResponse" }
        "400":
          description: Malformed payload (code "qr_malformed").
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "410": { $ref: "#/components/responses/Gone" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalError" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
...
  /api/v1/me/qr:
    post:
      operationId: issueQR
      tags: [Customer]
      summary: Mint a new one-time QR payload for identification at POS
      description: >
        The payload carries no PII and is one-time-use (see DEC-011).
        Re-issuing before the previous QR's issue cooldown elapses returns
        429 with Retry-After.
      security:
        - customerAccessToken: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IssueQRResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
...
  headers:
    RetryAfter:
      description: Seconds until the caller may retry.
      schema: { type: integer, minimum: 1 }
  responses:
...
    Unauthorized:
      description: Missing, invalid, or expired credential.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Credential valid but lacks the required scope/permission, or targets a resource outside its tenant.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Gone:
      description: >
        The referenced one-time resource is expired, already consumed, or
        never existed — deliberately a single response for all three (see
        DEC-011), so the caller can't distinguish them.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    TooManyRequests:
      description: Rate limit exceeded.
      headers:
        Retry-After: { $ref: "#/components/headers/RetryAfter" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ServiceUnavailable:
      description: A required backend (e.g. the rate limiter's Redis) is unreachable.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, description: Human-readable error message. }
        code:
          type: string
          description: >
            Machine-readable error code, present on a subset of error
            responses that callers may need to branch on programmatically
            (e.g. "daily_redeem_limit_exceeded", "qr_malformed",
            "over_refund"). Absent on errors with no defined code — callers
            must not assume this field is always present.
...
    IssueQRResponse:
      type: object
      properties:
        payload:
          type: string
          description: Opaque, versioned, one-time-use token. Contains no PII. The mobile app renders this as a QR image — the backend never generates image data.
        expires_at: { type: string, format: date-time }
        expires_in: { type: integer, description: "Seconds until expiry, at response time." }
    ResolveQRRequest:
      type: object
      required: [payload]
      properties:
        payload: { type: string }
    ResolveQRResponse:
      type: object
      properties:
        client_id: { type: integer, format: int64 }
        balance: { type: integer, format: int64 }
...
  securitySchemes:
    storeApiKey:
      type: http
      scheme: bearer
      description: >
        A per-store webhook API key (see Admin - API Keys), scoped to one
        or more of transactions.write / clients.lookup / balances.read. A
        store's legacy single-key credential (predating scopes) is treated
        as implicitly holding every scope.
    customerAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived customer access token from /auth/otp/verify or /auth/refresh.
    staffAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Short-lived staff access token from /staff/auth/oidc. Every
        operation under /admin also requires a specific RBAC permission —
        see domain.Permission and each operation's description — checked
        against the caller's active StaffMembership rows, resolved fresh on
        every request (never trusted from token claims).